Make it clearer that page example requires unique values
In this example, if the table is:
ID,Name,Age
0,Ted,5
1,James,5
2,Jill,5
3,Justin,30

and the page size = 1.
The first page will return '0,Ted,5' , but the second page will skip all other age=5 records and return '3,Justin,30'. 
So for this paging technique to be accurate, the paging property needs to be unique.  This PR makes that fact a bit clearer.