Make language edits
Both addon and add-ons was used in this file, and probably add-on with a hyphen is the better option to use. For example, see [add-on on Wikipedia](https://en.wikipedia.org/wiki/Add-on)

Converted NOTES to github flavour markdown notes, see [An option to highlight a "Note" and "Warning" using blockquote](https://github.com/orgs/community/discussions/16925)

Make various language edits.